feat: drag and drop
This is what the code looks like for building the entire thing in the drag/drop editor. Everything is 100% native except I had to add some code to align the social icons horizontally since the editor doesnt allow that.

Results are mixed. Looks okay on Outlook (still not perfect) but looks terrible on mobile. No control over how it styles for the smaller screen.